+++ plugin/lightroom-py-bridge.lrplugin/Info.lua
@@ -12,7 +12,7 @@ return {
 
   LrToolkitIdentifier = "com.henryshen.lightroom-py-bridge",
   LrPluginName = "lightroom-py bridge",
-  VERSION = { major = 0, minor = 4, revision = 0, build = 0 },
+  VERSION = { major = 0, minor = 4, revision = 2, build = 0 },
 
   LrPluginInfoUrl = "https://github.com/henryshen/lightroom-py",